chore(query+core+database+main): quick-sweep dead exports, fossils, and a misplaced module
Cleanup batch from the 2026-08-05 holistic sync-engine architecture review:

- query tests + core query-bindings.test import @wcpos/sync-engine/testing
  instead of reaching into the sibling package's src by relative path
  (testing door gains the two descriptor query-key exports)
- delete stableDescriptor identity fn (query-bindings.ts)
- query door drops dead exports: observeEngineCollections (now module-private),
  resetRefillRequirements, RequirementPlan
- move SyncStatusPersistenceBridge out of app/(app)/_layout.tsx into
  apps/main/lib/sync-status-persistence-bridge.tsx next to its test
- delete packages/database/src/plugins/example.ts (unused demo plugin with a
  live console.log)
- reset-collection.ts: retire emitCollectionReset + swappingCollections
  (no producer since the swapCollections path left @wcpos/query); reset$
  emission behavior unchanged
