+++ packages/database/src/plugins/reset-collection.ts
@@ -30,13 +30,6 @@ const removalCounts: Record<string, number> = {};
 const pendingReAdditions = new Set<string>();
 
 /**
- * Collections currently being swapped by swapCollections().
- * When a collection is in this set, the reset plugin will re-add it but NOT emit on reset$.
- * The swapCollections function handles the reset$ emission after the swap is complete.
- */
-export const swappingCollections = new Set<string>();
-
-/**
  * Set of collection names that this plugin manages.
  * Only these collections will be auto-recreated after removal.
  * Collections created by other plugins (e.g., FlexSearch) are excluded.
@@ -61,18 +54,6 @@ function isManagedCollection(collectionName: string, databaseName: string): bool
 // Subjects for emitting reset events
 const storeReset = new Subject<RxCollection>();
 const syncReset = new Subject<RxCollection>();
-
-/**
- * Manually emit a reset event for a collection.
- * Used by swapCollections after swap is complete.
- */
-export function emitCollectionReset(collection: RxCollection, databaseName: string): void {
-	if (databaseName.startsWith('fast_store')) {
-		syncReset.next(collection);
-	} else if (databaseName.startsWith('store')) {
-		storeReset.next(collection);
-	}
-}
 
 /**
  * Reset Collection Plugin
@@ -227,18 +208,10 @@ export const resetCollectionPlugin: RxPlugin = {
 
 						const cols = await database.addCollections({ [collectionName]: schema });
 
-						// Only emit on reset$ if NOT being swapped by swapCollections
-						// (swapCollections handles emission after swap completes)
-						if (!swappingCollections.has(collectionName)) {
-							syncReset.next(cols[collectionName]);
-							resetLogger.debug('Sync collection re-added and emitted reset$', {
-								context: { collection: collectionName },
-							});
-						} else {
-							resetLogger.debug('Sync collection re-added (no emit - in swap)', {
-								context: { collection: collectionName },
-							});
-						}
+						syncReset.next(cols[collectionName]);
+						resetLogger.debug('Sync collection re-added and emitted reset$', {
+							context: { collection: collectionName },
+						});
 					} else if (database.name.startsWith('store')) {
 						const schema = storeCollections[collectionName as keyof StoreCollections];
 						if (!schema) {
@@ -255,18 +228,10 @@ export const resetCollectionPlugin: RxPlugin = {
 
 						const cols = await database.addCollections({ [collectionName]: schema });
 
-						// Only emit on reset$ if NOT being swapped by swapCollections
-						// (swapCollections handles emission after swap completes)
-						if (!swappingCollections.has(collectionName)) {
-							storeReset.next(cols[collectionName]);
-							resetLogger.debug('Store collection re-added and emitted reset$', {
-								context: { collection: collectionName },
-							});
-						} else {
-							resetLogger.debug('Store collection re-added (no emit - in swap)', {
-								context: { collection: collectionName },
-							});
-						}
+						storeReset.next(cols[collectionName]);
+						resetLogger.debug('Store collection re-added and emitted reset$', {
+							context: { collection: collectionName },
+						});
 					}
 				} catch (error: any) {
 					resetLogger.error('Failed to re-add collection', {
